requirements.txt
django-environ
flask
pytube